+++ internal/storage/v1/badger/factory.go
@@ -17,12 +17,12 @@ import (
 	"github.com/spf13/viper"
 	"go.uber.org/zap"
 
+	depStore "github.com/jaegertracing/jaeger/internal/storage/v1/badger/dependencystore"
+	badgerSampling "github.com/jaegertracing/jaeger/internal/storage/v1/badger/samplingstore"
+	badgerStore "github.com/jaegertracing/jaeger/internal/storage/v1/badger/spanstore"
 	"github.com/jaegertracing/jaeger/pkg/distributedlock"
 	"github.com/jaegertracing/jaeger/pkg/metrics"
 	"github.com/jaegertracing/jaeger/plugin"
-	depStore "github.com/jaegertracing/jaeger/plugin/storage/badger/dependencystore"
-	badgerSampling "github.com/jaegertracing/jaeger/plugin/storage/badger/samplingstore"
-	badgerStore "github.com/jaegertracing/jaeger/plugin/storage/badger/spanstore"
 	"github.com/jaegertracing/jaeger/storage"
 	"github.com/jaegertracing/jaeger/storage/dependencystore"
 	"github.com/jaegertracing/jaeger/storage/samplingstore"
