+++ internal/storage/v1/badger/lock_test.go
