+++ internal/storage/v1/badger/options_test.go
