+++ internal/storage/v1/badger/spanstore/writer.go
