+++ internal/storage/v1/badger/stats_test.go
